support for building targets DS203, DS213, LA104 refactoring of shorter integer types, added override function attributes to prevent future errors, 64 bit architecture support (osx builds) some code cleanup stm32flash tool for flashing DS203

+++ system/os_host/source/framework/Classes.h
@@ -118,8 +118,7 @@ public:
 		return true;
 	}
 
-	// RAM optimization int->short
-	short left, top, right, bottom;
+	int left, top, right, bottom;
 };
 
 template <class TYPE>
